Finish newGrid button
I've fully built the system for creating a new
grid of a custom size. The button works and will
delete the old grid, and through the new listener
with a prompt you can pass a size into the old
function which I have changed to accept a specific
size. When I originally call the function I just
added a size to be passed.

I've also added an outline to the container and
deleted the cell outlines which I think looks
cleaner.