Reorganise page layout and add new instructions
I've increased the size of the grid to 800x800.
I've added new instructions to the left of the
grid that indicate how you use shade and regular
modes. And I've placed the colour options to the
right of the grid so as to increase the available
space to the grid so I could make it bigger.